agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
redis==5.1.0
supabase==2.9.0
PyGithub==2.4.0
python-dotenv==1.0.1
PyYAML==6.0.2
asyncio==3.4.3
websockets==13.1
sse-starlette==2.1.3
